Add automatic barcode generation and barcode printing

+++ sales/serializers.py
@@ -10,8 +10,9 @@ from .models import (
     SalesInvoice,
     SalesInvoiceItem,
     SalesRepresentative,
+    SalesPayment,
 )
-
+from django.db.models import Sum
 
 class SalesInvoiceItemSerializer(serializers.ModelSerializer):
 
@@ -26,6 +27,8 @@ class SalesInvoiceItemSerializer(serializers.ModelSerializer):
 
 
 class SalesInvoiceSerializer(serializers.ModelSerializer):
+    total_paid = serializers.SerializerMethodField()
+    remaining = serializers.SerializerMethodField()
 
     branch_name = serializers.CharField(
         source="branch.name",
@@ -59,6 +62,22 @@ class SalesInvoiceSerializer(serializers.ModelSerializer):
         model = SalesInvoice
         fields = "__all__"
 
+    def get_total_paid(self, obj):
+        return (
+            obj.payments.aggregate(
+                total=Sum("amount")
+            )["total"]
+            or 0
+        )
+
+    def get_remaining(self, obj):
+        paid = self.get_total_paid(obj)
+
+        return (
+            obj.total_amount_usd
+            - paid
+        )
+
     def validate(self, data):
         payment_type = data.get("payment_type")
         due_date = data.get("due_date")
@@ -82,34 +101,43 @@ class SalesInvoiceSerializer(serializers.ModelSerializer):
             item = item_data["item"]
             quantity = item_data["quantity"]
 
-            old_quantity = get_latest_quantity(
-                warehouse_id=invoice.warehouse.id,
-                item_id=item.id
-            )
-
-            if old_quantity < quantity:
-                invoice.delete()
-
-                raise ValidationError({
-                    "stock": f"Not enough stock for {item.name}. Available: {old_quantity}"
-                })
-
             invoice_item = SalesInvoiceItem.objects.create(
                 invoice=invoice,
                 **item_data
             )
 
-            new_quantity = old_quantity - quantity
+            if item.item_type != "service":
+                old_quantity = get_latest_quantity(
+                    warehouse_id=invoice.warehouse.id,
+                    item_id=item.id
+                )
 
-            Inventory.objects.create(
-                warehouse=invoice.warehouse,
-                item=item,
-                quantity=new_quantity,
-                operation_type="sale"
+                if old_quantity < quantity:
+                    invoice.delete()
+
+                    raise ValidationError({
+                        "stock": f"Not enough stock for {item.name}. Available: {old_quantity}"
+                    })
+
+                new_quantity = old_quantity - quantity
+
+                Inventory.objects.create(
+                    warehouse=invoice.warehouse,
+                    item=item,
+                    quantity=new_quantity,
+                    operation_type="sale"
+                )
+
+            total_usd += (
+                invoice_item.quantity
+                * invoice_item.unit_price_usd
             )
 
-            total_usd += invoice_item.quantity * invoice_item.unit_price_usd
-            total_syp += invoice_item.quantity * invoice_item.unit_price_syp
+            total_syp += (
+                invoice_item.quantity
+                * invoice_item.unit_price_syp
+            )
+
         invoice.total_amount_usd = total_usd
         invoice.total_amount_syp = total_syp
         invoice.total_amount = total_syp
@@ -117,9 +145,74 @@ class SalesInvoiceSerializer(serializers.ModelSerializer):
 
         return invoice
 
-
 class SalesRepresentativeSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = SalesRepresentative
         fields = "__all__"
+from django.db.models import Sum
+
+
+class SalesPaymentSerializer(
+    serializers.ModelSerializer
+):
+    invoice_number = serializers.SerializerMethodField()
+
+    customer = serializers.IntegerField(
+        source="invoice.customer.id",
+        read_only=True
+    )
+
+    customer_name = serializers.CharField(
+        source="invoice.customer.name",
+        read_only=True
+    )
+
+    invoice_total = serializers.DecimalField(
+        source="invoice.total_amount_usd",
+        max_digits=18,
+        decimal_places=2,
+        read_only=True
+    )
+    class Meta:
+        model = SalesPayment
+        fields = "__all__"
+    def get_invoice_number(self, obj):
+        return f"SI{str(obj.invoice.id).zfill(5)}"
+    def create(
+        self,
+        validated_data
+    ):
+
+        payment = SalesPayment.objects.create(
+            **validated_data
+        )
+
+        invoice = payment.invoice
+
+        total_paid = (
+            invoice.payments.aggregate(
+                total=Sum("amount")
+            )["total"]
+            or 0
+        )
+
+        invoice_total = (
+            invoice.total_amount_usd
+        )
+
+        if total_paid <= 0:
+
+            invoice.status = "unpaid"
+
+        elif total_paid < invoice_total:
+
+            invoice.status = "partial"
+
+        else:
+
+            invoice.status = "paid"
+
+        invoice.save()
+
+        return payment
